# api-definitions/ai-openapi.json
{
  "components": {
    "schemas": {
      "uk.gov.ons.addressIndex.model.server.response.rh.AddressByRHPartialAddressResponse": {
        "properties": {
          "input": {
            "type": "string"
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseAddressCustomRH"
            }
          },
          "filter": {
            "type": "string"
          },
          "fallback": {
            "type": "boolean"
          },
          "epoch": {
            "type": "string"
          },
          "limit": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "maxScore": {
            "type": "number",
            "format": "double"
          },
          "favourpaf": {
            "type": "boolean"
          },
          "favourwelsh": {
            "type": "boolean"
          },
          "countryBoosts": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.CountryBoosts"
          }
        },
        "required": [
          "input",
          "addresses",
          "filter",
          "fallback",
          "epoch",
          "limit",
          "offset",
          "total",
          "maxScore",
          "favourpaf",
          "favourwelsh",
          "countryBoosts"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseError": {
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.rh.AddressByRHPartialAddressResponseContainer": {
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "dataVersion": {
            "type": "string"
          },
          "response": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.rh.AddressByRHPartialAddressResponse"
          },
          "status": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseStatus"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseError"
            }
          }
        },
        "required": [
          "apiVersion",
          "dataVersion",
          "response",
          "status",
          "errors"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseAddressCustomRH": {
        "properties": {
          "uprn": {
            "type": "string"
          },
          "bestMatchAddress": {
            "type": "string"
          },
          "bestMatchAddressType": {
            "type": "string"
          },
          "censusAddressType": {
            "type": "string"
          },
          "censusEstabType": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          }
        },
        "required": [
          "uprn",
          "bestMatchAddress",
          "bestMatchAddressType",
          "censusAddressType",
          "censusEstabType",
          "countryCode"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseStatus": {
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.CountryBoosts": {
        "properties": {
          "eboost": {
            "type": "number",
            "format": "double"
          },
          "nboost": {
            "type": "number",
            "format": "double"
          },
          "sboost": {
            "type": "number",
            "format": "double"
          },
          "wboost": {
            "type": "number",
            "format": "double"
          },
          "lboost": {
            "type": "number",
            "format": "double"
          },
          "mboost": {
            "type": "number",
            "format": "double"
          },
          "jboost": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "eboost",
          "nboost",
          "sboost",
          "wboost",
          "lboost",
          "mboost",
          "jboost"
        ]
      },
...
    }
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/addresses/rh/partial": {
      "get": {
        "operationId": "partialAddressQueryRH",
        "tags": [
          "routes"
        ],
        "summary": "Search by partial address (for type ahead). RH version",
        "description": "Returns a list addresses, from a partial match on the input string",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "input",
            "description": "Specifies the address input.",
            "required": true
          },
          {
            "schema": {
              "default": false,
              "nullable": true,
              "type": "boolean"
            },
            "in": "query",
            "name": "fallback",
            "description": "Specifies whether a slow fallback query is used in the event of the main query returning no results.",
            "required": false
          },
          {
            "schema": {
              "default": 0,
              "nullable": true,
              "maximum": 250,
              "type": "integer"
            },
            "in": "query",
            "name": "offset",
            "description": "Specifies the offset from zero, used for pagination.",
            "required": false
          },
          {
            "schema": {
              "default": 10,
              "nullable": true,
              "maximum": 100,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "description": "Specifies the number of addresses to return.",
            "required": false
          },
          {
            "schema": {
              "nullable": true,
              "type": "string"
            },
            "in": "query",
            "name": "classificationfilter",
            "description": "Classification code filter. Can be pattern match (ZW*), exact match (RD06), multiple exact match (RD02,RD04) or a preset keyword such as residential, commercial, educational or workplace",
            "required": false
          },
          {
            "schema": {
              "nullable": true,
              "type": "string"
            },
            "in": "query",
            "name": "epoch",
            "description": "Select a specific AddressBase Epoch to search.",
            "required": false
          },
          {
            "schema": {
              "default": true,
              "nullable": true,
              "type": "boolean"
            },
            "in": "query",
            "name": "favourpaf",
            "description": "paf beats nag on draw for best match",
            "required": false
          },
          {
            "schema": {
              "default": false,
              "nullable": true,
              "type": "boolean"
            },
            "in": "query",
            "name": "favourwelsh",
            "description": "Welsh beats English on draw for best match",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "eboost",
            "description": "Weighting for addresses in England as a decimal from 0 to 10",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "nboost",
            "description": "Weighting for addresses in Northern Ireland as a decimal from 0 to 10",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "sboost",
            "description": "Weighting for addresses in Scotland as a decimal from 0 to 10",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "wboost",
            "description": "Weighting for addresses in Wales as a decimal from 0 to 10",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "lboost",
            "description": "Weighting for addresses in Channel Islands as a decimal from 0 to 10",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "mboost",
            "description": "Weighting for addresses in Isle of Man as a decimal from 0 to 10",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "jboost",
            "description": "Weighting for addresses outside country boundaries as a decimal from 0 to 10",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Success. A json return of matched addresses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.rh.AddressByRHPartialAddressResponseContainer"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Indicates an issue with the request. Further details are provided in the response."
          },
          "401": {
            "description": "Unauthorised. The API key provided with the request is invalid."
          },
          "429": {
            "description": "Server too busy. The Address Index API is experiencing exceptional load."
          },
          "500": {
            "description": "Internal server error. Failed to process the request due to an internal error."
          }
        }
      }
    },
...
  },
  "info": {
    "title": "ONS Address Index API",
    "description": "The ONS Address Index API lets you retrieve the official version of addresses in England and Wales. The data is currently updated on a 6 week rolling basis.",
    "version": "see /version below"
  },
...
}